Login is needed to use apps' features

+++ src/php/bookInfo.php
@@ -1,13 +1,20 @@
 <?php
 
-//PHP proxy file to pass in ISBN and return the JSON from is ISBNdb website
-$isbn = $_GET['isbn'];
-$filename = "./" . $isbn . "json";
+session_start();
 
-    file_put_contents($filename, file_get_contents("http://isbndb.com/api/v2/json/XBOU017W/book/" . $isbn));
+if (!isset($_SESSION['id']))
+	echo "err-notloggedin";
+else
+{
+	//PHP proxy file to pass in ISBN and return the JSON from is ISBNdb website
+	$isbn = $_GET['isbn'];
+	$filename = "./" . $isbn . "json";
 
-header('Content-Type: application/json; charset=utf-8');
+		file_put_contents($filename, file_get_contents("http://isbndb.com/api/v2/json/XBOU017W/book/" . $isbn));
 
-echo (file_get_contents($filename));
-unlink($filename);
+	header('Content-Type: application/json; charset=utf-8');
+
+	echo (file_get_contents($filename));
+	unlink($filename);
+}
 ?>
